Share GEventEvaluator with logback-access
