+++ logback-access/src/main/java/ch/qos/logback/access/boolex/GEventEvaluator.java
@@ -1,0 +1,6 @@
+package ch.qos.logback.access.boolex;
+
+import ch.qos.logback.access.spi.IAccessEvent;
+import ch.qos.logback.core.boolex.GEventEvaluatorBase;
+
+public class GEventEvaluator extends GEventEvaluatorBase<IAccessEvent> {}
